// src/app/main/shippings/models/carrier-tracking-numbers-group.model.ts
export class CarrierTrackingNumbersGroup {
    availNumbers?: number;
    
    carrierId?: number;
    carrierName?: string;
    carrierCode?: string;

    branchId?: number;
    branchName?: string;
    branchCode?: string;    

    companyGroupId?: number;
    companyGroupName?: string;

    id?: number;
}